pyproject.toml
[project]
name = "pycharm"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.51.0",
    "arxiv>=2.2.0",
    "matplotlib>=3.10.3",
    "mcp>=1.9.0",
    "mcp-server-fetch>=2025.4.7",
    "notebook>=7.4.2",
    "numpy>=2.2.6",
    "openai>=1.79.0",
    "pandas>=2.2.3",
    "pydantic>=2.11.4",
    "pydantic-ai>=0.2.4",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.6.1",
    "torch>=2.7.0",
]
